server/requirements.txt
flask
python-dotenv
flask-restx
flask-sqlalchemy
flask-cors
pgvector
Flask-Login
Werkzeug